FB-51 : Add Status & Visibility To Event Entity
STATUS will be either: LIVE or CANCELED. For now this won’t do anything but will be used in future for the cancel event endpoint. It’ll be used to not show the event if it’s canceled. (edited)
VISIBILITY: Will be either PUBLIC or PRIVATE. This will be used to fetermine if it should show up in search or getAllEvents endpoints. This one can be edited via Postman.

Unit tests added.